refactor: post source info
